+++ src/components/Home/Products/SignleMainProducts/SingleProduct/Product_Detail/ProductDetails.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ProductDetails = () => {
+  return <div className="w-full h-full ">ProductDetails</div>;
+};
+
+export { ProductDetails };
